Fix Vue frontend
