+++ jsonconfig.json
@@ -1,0 +1,5 @@
+{
+  "include": [
+    "./src/main/web/**/*"
+  ]
+}
